+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Replace types like `Tuple2<String, int>` with `(String, int)`.
   * Replace constructors like `Tuple2('hello', 42)` with `('hello', 42)`.
 * Add `Iterable.pairwise` over the successive overlapping pairs.
+* Add `zip` extension that combines a tuple of iterables to an iterable of tuples.
 
 ## 3.10.0
 * Add a growable `BitList`, make it behave more like `List<bool>`.